Expanded backup and requeue code to backup (but not requeue) right before a group finding step, and to backup (but not requeue) halfway through the job. The NVME burst buffer auto-IO is, apparently, still under development at Oak Ridge, but I'm in touch with them about it. For now, all the backups are done in a blocking manner directly to the global disk. Modified auto-requeue to not burn up remaining time left in job but immediately requeue and kill the job. Also removed a minor redundancy in cosmo_setup.cpp.

+++ Abacus/abacus.py
@@ -58,9 +58,8 @@ from . import GenParam
 from . import zeldovich
 from Abacus.Cosmology import AbacusCosmo
 
-
+NEEDS_INTERIM_BACKUP_MINS = 240 #minimum job runtime (in minutes) for which we'd like to do a backup halfway through the job. 
 EXIT_REQUEUE = 200
-RUN_TIME_MINUTES = os.getenv("JOB_ACTION_WARNING_TIME")
 
 site_param_fn = pjoin(abacuspath, 'Production', 'site_files', 'site.def')
 directory_param_fn = pjoin(abacuspath, 'Abacus', 'directory.def')
@@ -128,7 +127,7 @@ def run(parfn='abacus.par2', config_dir=path.curdir, maxsteps=10000, clean=False
     icdir = params['InitialConditionsDirectory']
     outdir = params['OutputDirectory']
     logdir = params['LogDirectory']
-    groupdir = params.get('GroupDirectory', '')
+    groupdir  = params.get('GroupDirectory', '')
     basedir = params['WorkingDirectory']
 
     if parallel:
@@ -715,19 +714,6 @@ class StatusLogWriter:
         self.log_fp.write(('\n * ' + fmtstring.format(*args, **kwargs) + end).encode('utf-8'))
         self.log_fp.flush()
     
-
-
-
-class SignalHandler(object):
-    def __init__(self):
-        self.graceful_exit = None
-    def handle(self, signalNumber, frame):
-        print('Received signal:', signalNumber)
-        self.graceful_exit = 1
-        print('We are running out of time in the job! Setting signal handler graceful_exit flag to ', self.graceful_exit)
-
- 
- 
  
 def singlestep(paramfn, maxsteps=None, make_ic=False, stopbefore=-1, resume_dir=None):
     """
@@ -758,11 +744,6 @@ def singlestep(paramfn, maxsteps=None, make_ic=False, stopbefore=-1, resume_dir=
         or EXIT_REQUEUE if we finish cleanly but have not reached the final z
     """
     
-    s = SignalHandler()
-    signal.signal(signal.SIGUSR1, s.handle)
-    signal.signal(signal.SIGUSR2, s.handle)
-    
-    
     finished = False
 
     if maxsteps is None:
@@ -787,7 +768,7 @@ def singlestep(paramfn, maxsteps=None, make_ic=False, stopbefore=-1, resume_dir=
     
     if parallel:
         # TODO: figure out how to signal a backup to the nodes
-        run_time_minutes = int(RUN_TIME_MINUTES) #360
+        run_time_minutes = int(os.getenv("JOB_ACTION_WARNING_TIME"))
         run_time_secs = 60 * run_time_minutes
         start_time = wall_timer()
         print("Beginning run at time", start_time, ", running for ", run_time_minutes, " minutes.\n")
@@ -847,30 +828,24 @@ def singlestep(paramfn, maxsteps=None, make_ic=False, stopbefore=-1, resume_dir=
         
         if not make_ic: #if this is not a clean run, redistribute the state out to the nodes. 
             print('Distributing in order to resume...')
-            distribute_state_cmd = [pjoin(abacuspath, 'Abacus', 'move_node_states.py'), paramfn, '--distribute', resume_dir,  '--verbose']            
+            distribute_state_cmd = [pjoin(abacuspath, 'Abacus', 'move_node_states.py'), paramfn, '--distribute', resume_dir]            
             distribute_fns_present = subprocess.check_call(Conv_mpirun_cmd + distribute_state_cmd)
             
-            
-#            if not distribute_fns_present: 
-#                raise RuntimeError('"Missing/corrupted files detected during distribute to resume. Exiting!')
-#                #Build the recover_multipoles executable
-#                 with Tools.chdir(pjoin(abacuspath, "singlestep")):
-#                     subprocess.check_call(['make', 'recover_multipoles'])
-#
-#                 reconstruct_read_multipoles_cmd = [pjoin(abacuspath, 'singlestep', 'recover_multipoles'), paramfn]
    
-   
-    emergency_exit_fn = pjoin(param['WorkingDirectory'], 'abandon_ship')
     print("Beginning abacus steps:")
-    print("\n------------------")
-    print("To trigger emergency quit safely, create file", emergency_exit_fn)
-    print("------------------")
+    if parallel:
+        emergency_exit_fn = pjoin(param['WorkingDirectory'], 'abandon_ship')
+        print("\n------------------")
+        print("To trigger emergency quit safely, create file", emergency_exit_fn)
+        print("------------------")
     
     singlestep_cmd = [pjoin(abacuspath, "singlestep", "singlestep"), paramfn, str(int(make_ic))]
     if parallel:
         singlestep_cmd = mpirun_cmd + singlestep_cmd
-        print("Using singlestep_cmd ", singlestep_cmd)
-        
+        print("Using singlestep_cmd ", singlestep_cmd) 
+    
+    #if this job is longer than a set time (NEEDS_INTERIM_BACKUP_MINS), we'll need to do a backup halfway through the run. 
+    interim_backup_complete = run_time_minutes <= NEEDS_INTERIM_BACKUP_MINS
 
     for i in range(maxsteps):
         
@@ -1036,31 +1011,69 @@ def singlestep(paramfn, maxsteps=None, make_ic=False, stopbefore=-1, resume_dir=
                 np.savez(pjoin(param.LogDirectory, f"step{write_state.FullStepNumber:04d}.pow"), k=k,P=P,nb=nb)
                 os.remove(dfn)
         
-        abandon_ship = path.exists(emergency_exit_fn)
-        graceful_exit = (wall_timer() - start_time >= run_time_secs) or abandon_ship
+        if parallel:
+            # check if we need to gracefully exit and/or backup:
+            #    is a group finding step next?          ---> backup to NVME and continue.   (group finding steps are more likely to crash!)
+            #    are we ~halfway through our long run?  ---> backup to NVME and continue.   (just in case.)
+            #    are we running out of time in the job? ---> backup to global dir and exit. 
+            #    did the user ask to abandon ship?      ---> backup to global dir and exit. 
+             
+            
+            abandon_ship = path.exists(emergency_exit_fn)
+            out_of_time = (wall_timer() - start_time >= run_time_secs)
+            
+            #if the halfway-there backup hasn't been done already, and we're more than halfway through the job, backup the state now: 
+            interim_backup = (not interim_backup_complete) and (wall_timer() - start_time >= run_time_secs / 2)
+            
+            #are we coming up on a group finding step? If yes, backup the state, just in case. 
+            pre_gf_backup  = False 
+            nGFoutputs = len(param.L1OutputRedshifts)
+            for i in range(nGFoutputs):
+                L1z = param.L1OutputRedshifts[i] 
+                if L1z <= -1:
+                    continue
+                L1a = 1.0/(1.0+L1z)                
+                
+                #here we assume that the next da will be similar to the previous one. 
+                if (write_state.Redshift > L1z + 1e-12) and ( 1.0/(1.0+write_state.Redshift) + write_state.DeltaScaleFactor > L1a ):
+                    pre_gf_backup = True 
+            
+            exit = out_of_time or abandon_ship
+            save = exit or interim_backup or pre_gf_backup
         
-        if parallel and graceful_exit:
-            print("Current time: ", wall_timer(), start_time, run_time_secs)
-            restore_time = wall_timer()
-            
-            if abandon_ship:
-                print('\nAbandoning ship!\n')
+            if save:
+                if abandon_ship:
+                    exit_message = 'EMERGENCY EXIT REQUESTED: '
+                if out_of_time:
+                    exit_message = 'RUNNING OUT OF JOB TIME: '                    
+                if interim_backup:
+                    exit_message = 'HALFWAY THROUGH A LONG JOB: '                    
+                if pre_gf_backup:
+                    exit_message = 'GROUP FINDING COMING UP: '
+                    
+                print(exit_message, 'backing up node state.')
                 
-
-            print('Graceful exit triggered. Retrieving state from nodes and saving in global directory.')
-            retrieve_state_cmd = [pjoin(abacuspath, 'Abacus', 'move_node_states.py'), paramfn, resume_dir, '--retrieve', '--verbose']
-            subprocess.check_call(Conv_mpirun_cmd + retrieve_state_cmd)
+                restore_time = wall_timer()
+        
+                retrieve_state_cmd = [pjoin(abacuspath, 'Abacus', 'move_node_states.py'), paramfn, resume_dir, '--retrieve']
+                subprocess.check_call(Conv_mpirun_cmd + retrieve_state_cmd)
             
-            restore_time = wall_timer() - restore_time
+                restore_time = wall_timer() - restore_time
             
-            print(f'Retrieving and storing state took {restore_time} seconds\n')
-            print('Exiting and requeueing.')
+                print(f'Retrieving and storing state took {restore_time} seconds. ', end = '')
             
-            #checking if path exists explicitly just in case user requested emergency exit while we were retrieveing the state. 
-            if path.exists(emergency_exit_fn): 
-                os.remove(emergency_exit_fn)
+                #checking if path exists explicitly just in case user requested emergency exit while we were retrieveing the state. 
+                if path.exists(emergency_exit_fn): 
+                    os.remove(emergency_exit_fn)
                 
-            return EXIT_REQUEUE  
+                if interim_backup:
+                    interim_backup_complete = True 
+                    
+                if exit:
+                    print('Exiting and requeueing.')
+                    return EXIT_REQUEUE  
+                else:
+                    print('Continuing run.')
         
         # Now shift the states down by one
         move_state_dirs(read, write, past)
@@ -1088,7 +1101,7 @@ def singlestep(paramfn, maxsteps=None, make_ic=False, stopbefore=-1, resume_dir=
             finished = True
             break 
             
-        if abandon_ship:
+        if parallel and abandon_ship:
             print(f"Abandon ship triggered! Terminating job.")
             os.remove(emergency_exit_fn)
             break
